+++ app/org/nlp4l/framework/actors/ProcessorChainActor.scala
@@ -22,7 +22,7 @@ import org.nlp4l.framework.processors.ProcessorChain2
 import akka.actor.Actor
 import akka.actor.Props
 import javax.inject.Inject
-import org.nlp4l.framework.buildin.JobMessage
+import org.nlp4l.framework.builtin.JobMessage
 
 
 class ProcessorChain2Actor @Inject()(jobDAO: JobDAO, runDAO: RunDAO) extends Actor {
