+++ app/org/nlp4l/framework/builtin/DeployerChain.scala
@@ -14,78 +14,69 @@
  * limitations under the License.
  */
 
-package org.nlp4l.framework.buildin
+package org.nlp4l.framework.builtin
 
 import scala.collection.mutable
 import scala.concurrent.Await
 import scala.collection.convert.WrapAsScala._
 
+import org.joda.time.DateTime
 import org.nlp4l.framework.dao.JobDAO
 import org.nlp4l.framework.models.Dictionary
-import org.nlp4l.framework.processors.Validator
-import org.nlp4l.framework.processors.ValidatorFactory
+import org.nlp4l.framework.processors.Deployer
+import org.nlp4l.framework.processors.DeployerFactory
 
 import com.typesafe.config.ConfigFactory
 
 import play.api.Logger
-import play.api.libs.concurrent.Execution.Implicits.defaultContext
 
-class ValidatorChain (val chain: List[Validator]) {
+class DeployerChain (val chain: List[Deployer]) {
   private val logger = Logger(this.getClass)
   
-  def process(dic: Dictionary): Seq[String] = {
+  def process(jobDAO: JobDAO, jobId: Int, runId: Int, dic: Dictionary): Seq[String] = {
     var errMsg: Seq[String] = Seq()
-    def loop(li: List[Validator], data:Option[Dictionary] = None): Unit = li match {
+    def loop(li: List[Deployer], data:Option[Dictionary] = None): Unit = li match {
       case Nil => ()
       case head :: Nil =>
-        val out: Tuple2[Boolean, Seq[String]] = head.validate(data)
+        val out: Tuple2[Boolean, Seq[String]] = head.deploy(data)
         if(!out._1) errMsg = errMsg union out._2
       case head :: tail =>
-        val out: Tuple2[Boolean, Seq[String]] = head.validate(data)
+        val out: Tuple2[Boolean, Seq[String]] = head.deploy(data)
         if(!out._1) errMsg = errMsg union out._2
         loop(tail, data)
     }
     loop(chain, Some(dic))
+    val job = Await.result(jobDAO.get(jobId), scala.concurrent.duration.Duration.Inf)
+    jobDAO.update(Job(job.jobId, job.name, job.config, runId, job.lastRunAt, Some(new DateTime())))
     errMsg
   }
 }
 
-object ValidatorChain {
+object DeployerChain {
 
-  // Processor
-  private var mapP: Map[Int, ValidatorChain] = null
-  def chainMap: Map[Int, ValidatorChain] = mapP
-  
-  def loadChain(jobDAO: JobDAO, jobId: Int): Unit = {
-    jobDAO.get(jobId).map(
-        job => 
-           mapP += (jobId -> new ValidatorChainBuilder().build(job.config).result())
-    )
-  }
-  
-  def getChain(jobDAO: JobDAO, jobId: Int): ValidatorChain = {
+  def getChain(jobDAO: JobDAO, jobId: Int): DeployerChain = {
     val job = Await.result(jobDAO.get(jobId), scala.concurrent.duration.Duration.Inf)
-    new ValidatorChainBuilder().build(job.config).result()
+    new DeployerChainBuilder().build(job.config).result()
   }
 }
 
 
-class ValidatorChainBuilder() {
+class DeployerChainBuilder() {
   val logger = Logger(this.getClass)
-  val buf = mutable.ArrayBuffer[Validator]()
+  val buf = mutable.ArrayBuffer[Deployer]()
 
-  def build(confStr: String): ValidatorChainBuilder = {
+  def build(confStr: String): DeployerChainBuilder = {
     val config = ConfigFactory.parseString(confStr)
 
-    val v = config.getConfigList("validators")
+    val v = config.getConfigList("deployers")
     v.foreach {
       pConf =>
         try {
           val className = pConf.getString("class")
           val constructor = Class.forName(className).getConstructor(classOf[Map[String, String]])
           val settings = pConf.getConfig("settings").entrySet().map(f => f.getKey -> f.getValue.unwrapped()).toMap
-          val facP = constructor.newInstance(settings).asInstanceOf[ValidatorFactory]
-          val p:Validator = facP.getInstance()
+          val facP = constructor.newInstance(settings).asInstanceOf[DeployerFactory]
+          val p:Deployer = facP.getInstance()
           buf += p
         } catch {
           case e: Exception => logger.error(e.getMessage)
@@ -94,5 +85,5 @@ class ValidatorChainBuilder() {
     this
   }
 
-  def result() = new ValidatorChain(buf.toList)
+  def result() = new DeployerChain(buf.toList)
 }
