+++ app/org/nlp4l/framework/builtin/ValidatorChain.scala
@@ -14,69 +14,78 @@
  * limitations under the License.
  */
 
-package org.nlp4l.framework.buildin
+package org.nlp4l.framework.builtin
 
 import scala.collection.mutable
 import scala.concurrent.Await
 import scala.collection.convert.WrapAsScala._
 
-import org.joda.time.DateTime
 import org.nlp4l.framework.dao.JobDAO
 import org.nlp4l.framework.models.Dictionary
-import org.nlp4l.framework.processors.Deployer
-import org.nlp4l.framework.processors.DeployerFactory
+import org.nlp4l.framework.processors.Validator
+import org.nlp4l.framework.processors.ValidatorFactory
 
 import com.typesafe.config.ConfigFactory
 
 import play.api.Logger
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
 
-class DeployerChain (val chain: List[Deployer]) {
+class ValidatorChain (val chain: List[Validator]) {
   private val logger = Logger(this.getClass)
   
-  def process(jobDAO: JobDAO, jobId: Int, runId: Int, dic: Dictionary): Seq[String] = {
+  def process(dic: Dictionary): Seq[String] = {
     var errMsg: Seq[String] = Seq()
-    def loop(li: List[Deployer], data:Option[Dictionary] = None): Unit = li match {
+    def loop(li: List[Validator], data:Option[Dictionary] = None): Unit = li match {
       case Nil => ()
       case head :: Nil =>
-        val out: Tuple2[Boolean, Seq[String]] = head.deploy(data)
+        val out: Tuple2[Boolean, Seq[String]] = head.validate(data)
         if(!out._1) errMsg = errMsg union out._2
       case head :: tail =>
-        val out: Tuple2[Boolean, Seq[String]] = head.deploy(data)
+        val out: Tuple2[Boolean, Seq[String]] = head.validate(data)
         if(!out._1) errMsg = errMsg union out._2
         loop(tail, data)
     }
     loop(chain, Some(dic))
-    val job = Await.result(jobDAO.get(jobId), scala.concurrent.duration.Duration.Inf)
-    jobDAO.update(Job(job.jobId, job.name, job.config, runId, job.lastRunAt, Some(new DateTime())))
     errMsg
   }
 }
 
-object DeployerChain {
+object ValidatorChain {
 
-  def getChain(jobDAO: JobDAO, jobId: Int): DeployerChain = {
+  // Processor
+  private var mapP: Map[Int, ValidatorChain] = null
+  def chainMap: Map[Int, ValidatorChain] = mapP
+  
+  def loadChain(jobDAO: JobDAO, jobId: Int): Unit = {
+    jobDAO.get(jobId).map(
+        job => 
+           mapP += (jobId -> new ValidatorChainBuilder().build(job.config).result())
+    )
+  }
+  
+  def getChain(jobDAO: JobDAO, jobId: Int): ValidatorChain = {
     val job = Await.result(jobDAO.get(jobId), scala.concurrent.duration.Duration.Inf)
-    new DeployerChainBuilder().build(job.config).result()
+    new ValidatorChainBuilder().build(job.config).result()
   }
 }
 
 
-class DeployerChainBuilder() {
+class ValidatorChainBuilder() {
   val logger = Logger(this.getClass)
-  val buf = mutable.ArrayBuffer[Deployer]()
+  val buf = mutable.ArrayBuffer[Validator]()
 
-  def build(confStr: String): DeployerChainBuilder = {
+  def build(confStr: String): ValidatorChainBuilder = {
     val config = ConfigFactory.parseString(confStr)
 
-    val v = config.getConfigList("deployers")
+    val v = config.getConfigList("validators")
     v.foreach {
       pConf =>
         try {
           val className = pConf.getString("class")
           val constructor = Class.forName(className).getConstructor(classOf[Map[String, String]])
           val settings = pConf.getConfig("settings").entrySet().map(f => f.getKey -> f.getValue.unwrapped()).toMap
-          val facP = constructor.newInstance(settings).asInstanceOf[DeployerFactory]
-          val p:Deployer = facP.getInstance()
+          val facP = constructor.newInstance(settings).asInstanceOf[ValidatorFactory]
+          val p:Validator = facP.getInstance()
           buf += p
         } catch {
           case e: Exception => logger.error(e.getMessage)
@@ -85,5 +94,5 @@ class DeployerChainBuilder() {
     this
   }
 
-  def result() = new DeployerChain(buf.toList)
+  def result() = new ValidatorChain(buf.toList)
 }
