+++ app/org/nlp4l/framework/dao/RunDao.scala
@@ -31,12 +31,12 @@ import org.joda.time.DateTime
 import org.nlp4l.framework.models.Cell
 import org.nlp4l.framework.models.CellAttribute
 import org.nlp4l.framework.models.CellType
-import org.nlp4l.framework.buildin.DbModels.resultAsStringMap
+import org.nlp4l.framework.builtin.DbModels.resultAsStringMap
 import org.nlp4l.framework.models.Dictionary
 import org.nlp4l.framework.models.DictionaryAttribute
 import org.nlp4l.framework.models.Record
-import org.nlp4l.framework.buildin.Job
-import org.nlp4l.framework.buildin.JobStatus
+import org.nlp4l.framework.builtin.Job
+import org.nlp4l.framework.builtin.JobStatus
 
 class RunDAO @Inject()(protected val dbConfigProvider: DatabaseConfigProvider) extends HasDatabaseConfigProvider[JdbcProfile] { 
   import driver.api._
