+++ app/org/nlp4l/framework/views/editjob.scala.html
@@ -1,4 +1,4 @@
-@import org.nlp4l.framework.buildin.Job
+@import org.nlp4l.framework.builtin.Job
 @(job: Job, runIdList: Seq[Int], success: String, error: String)
 
 @headerAddon = {
